Split up the profile screen composable
