--- feature/profile/src/commonMain/kotlin/com/chesire/nekomp/feature/profile/ui/ViewEvent.kt
@@ -1,3 +1,0 @@
-package com.chesire.nekomp.feature.profile.ui
-
-sealed interface ViewEvent
